pyproject.toml
[project]
name = "azure-cost-analyzer"
version = "0.1.0"
description = "Developers often overspend on cloud; this tool: Fetches Azure VM, app service, DB cost Predicts monthly bill Shows per-service breakdown Alerts if budget exceeds a limit Suggests cost-optimization tips"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "azure-identity>=1.25.1",
    "azure-mgmt-costmanagement>=4.0.1",
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "loguru>=0.7.3",
    "pathlib>=1.0.1",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "pyupgrade>=3.21.2",
    "ruff>=0.14.13",
    "ty>=0.0.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
